figma-transformer: split site generation contracts
